Updated relaxation and performance
The use of ^ contributes a large portion of the run time. Replace it
with wrapper to make use of tranformation to * operations.
This can lead to a 10-fold improvement in performance.

Also added @inbounds when collecting grandients to optimise the
performance.

Added options to limit stepsize in LBFGS during line search