Critical bug fix for multi-specie calculation
Fix a critical bug for multi-specie support, where the calculation
of feature vectors and gradients of the feature vectors are incorrect.
The issue is that the buffer for storing gradient was not cleared/
accumulated correctly during the calculation. While the feature vector
updates are "global", the gradients calculated are only for specific
pair/triplets of ijk.