Performance optimisations
Added threaded summation of forces. For large systems, the cost
apply the chain rule for forces becomes significant. It might be
useful to use a two-pass approach instead.

In the first pass, compute the feature vectors only, then pass the
vectors to the NN for gradients. Then a second pass can be conducted.
The second pass should cost minimum overhead as distance vectors have
already being built, apart from vjk, but the cost of it should be
minimum compared to computing the gradient itself.