Update the multiple specie feature matching
Enforce the matching of the first item, which ensures an atomic-centric
view. Previously, :H:O:O is equivalent to :O:H:O, but it is not the
case any more. However, :O:H:O and :O:O:H are still equivalent, e.g.
the permutation matching only considers those beyond the first item.