Added TPSD optimisation
As an alternative to the Optim interface, which has a tendency to
crash the calculation by shrinking the cell too much? The alternative
is to implement TPSD as a Optim optimiser....
Keeping the two interface for now.

Also updated the thread code eliminating the assumping that threadid()
will not change during the execution. This cannot be gauranteed with
the dynamic scheduler.