Apply formatting
Algorithm works for feature generation without gradient.
Aim for maximum efficiency when p==q==rcut==f so all features are
computed with the same power, cut off and function. This may not be
compatible with cases with specie-dependent cut-off/power to further
simplify the model, in which case more complex logic needs to be
implemented.